backend/requirements.txt
Flask
PyPDF2
python-docx
nltk
scikit-learn